pyproject.toml
[tool.poetry]
name = "md2site"
version = "0.0.5"
description = "A static site generator with bare-minimum functionality."
license = "GPL-3.0-or-later"
authors = ["Yoonseop Kang <e0engoon@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/e0en/md2site"
repository = "https://github.com/e0en/md2site"
classifiers = [
  "Development Status :: 1 - Planning",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python :: 3",
  "Topic :: Text Processing :: Markup :: Markdown",
  "Topic :: Internet",
  "Typing :: Typed",
]

[tool.poetry.scripts]
md2site = "md2site.generator:generate"

[tool.poetry.dependencies]
python = "^3.10.5"
Jinja2 = "^3.1.2"
toml = "^0.10.2"
mistletoe = "^1.1.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pre-commit = "^3.3.3"
ruff = "^0.0.275"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
